+++ client-mr/src/main/java/org/apache/hadoop/mapred/SortWriteBufferManager.java
@@ -240,7 +240,7 @@ public class SortWriteBufferManager<K, V> {
           for (ShuffleBlockInfo block : shuffleBlocks) {
             size += block.getFreeMemory();
           }
-          SendShuffleDataResult result = shuffleWriteClient.sendShuffleData(appId, shuffleBlocks, () -> true);
+          SendShuffleDataResult result = shuffleWriteClient.sendShuffleData(appId, shuffleBlocks, () -> false);
           successBlockIds.addAll(result.getSuccessBlockIds());
           failedBlockIds.addAll(result.getFailedBlockIds());
         } catch (Throwable t) {
