+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -105,7 +105,7 @@ public class RssShuffleManager implements ShuffleManager {
         SendShuffleDataResult result = shuffleWriteClient.sendShuffleData(
             appId,
             shuffleDataInfoList,
-            () -> isValidTask(taskId)
+            () -> !isValidTask(taskId)
         );
         putBlockId(taskToSuccessBlockIds, taskId, result.getSuccessBlockIds());
         putBlockId(taskToFailedBlockIds, taskId, result.getFailedBlockIds());
