+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/writer/RssShuffleWriter.java
@@ -164,14 +164,14 @@ public class RssShuffleWriter<K, V, C> extends ShuffleWriter<K, V> {
   @Override
   public void write(Iterator<Product2<K, V>> records) {
     try {
-      doWrite(records);
+      writeImpl(records);
     } catch (Exception e) {
       taskFailureCallback.apply(taskId);
       throw e;
     }
   }
 
-  private void doWrite(Iterator<Product2<K,V>> records) {
+  private void writeImpl(Iterator<Product2<K,V>> records) {
     List<ShuffleBlockInfo> shuffleBlockInfos = null;
     Set<Long> blockIds = Sets.newConcurrentHashSet();
     while (records.hasNext()) {
